Initial commit of work.
- Complete "calculate_rank" file that will take in match results file and create rankings file.
- Initial unit tests to check the "points" and "ordering" functions.
- Some infrastructure files.... not many
- Dockefile to run this anywhere - assuming docker is installed.
- requirements.txt to list the expected python packages we need to run our tests.

The code itselfs seems to work as required, and we have covered the majority of the files with tests.
We still need a test to confirm that our "calculate" function runs as expected.
We also need to change the README file with details on how to run the tests, and use docker, if required.